LayoutTests/fast/js/sputnik/Conformance/12_Statement/12.11_The_switch_Statement/S12.11_A3_T1-expected.txt

CONSOLE MESSAGE: line 79: SyntaxError: Parse error
S12.11_A3_T1

PASS Expected parsing failure

TEST COMPLETE
